// src/app/pages/budgets/components/budgets-table/budgets.interface.ts
export interface BudgetsInterface {
    nickname: string;
    usuarioauth: string;
    claveauth: string;
    idempresa: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    costo: number;
    idstatususuario: number;
    emailsms: string;
    bfechainicial: boolean;
    fechainicial: string;
}